refactor: split up signup form into three forms
